@@ -20,14 +20,17 @@ import android.content.Context;
 import android.media.AudioManager;
 import android.media.MediaPlayer;
 import android.os.Bundle;
-import android.support.v4.app.Fragment;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.AdapterView;
 import android.widget.ListView;
 
+import androidx.fragment.app.Fragment;
+
 import java.util.ArrayList;
+
+// import android.support.v4.app.Fragment;
 
 /**
  * {@link Fragment} that displays a list of phrases.
